Break down stack.py into smaller, more modular classes
