+++ source/tests/integration/infrastructure/test_buckets.py
@@ -7,7 +7,7 @@ import os
 import typing
 import boto3
 
-from refreezer.infrastructure.stack import OutputKeys
+from refreezer.infrastructure.output_keys import OutputKeys
 
 if typing.TYPE_CHECKING:
     from mypy_boto3_s3 import S3Client
