+++ source/tests/integration/infrastructure/test_chunk_retrieval_lambda.py
@@ -8,7 +8,7 @@ import typing
 import boto3
 import json
 
-from refreezer.infrastructure.stack import OutputKeys
+from refreezer.infrastructure.output_keys import OutputKeys
 
 if typing.TYPE_CHECKING:
     from mypy_boto3_lambda import LambdaClient
