+++ source/tests/integration/infrastructure/test_retrieve_archive_step_function.py
@@ -6,7 +6,7 @@ SPDX-License-Identifier: Apache-2.0
 import os
 import typing
 import boto3
-from refreezer.infrastructure.stack import OutputKeys
+from refreezer.infrastructure.output_keys import OutputKeys
 
 if typing.TYPE_CHECKING:
     from mypy_boto3_stepfunctions import SFNClient
